// OlymppMarketsClient/src/app/models/product.model.ts
export interface Product {
    id?: number;         // Identifiant du produit
    name: string;        // Nom du produit
    brand: string;       // Marque du produit
    size: string;        // Taille du produit
    price: number;       // Prix du produit
    stock: number;       // Stock disponible du produit
  }
  